+++ .gitignore
@@ -10,5 +10,6 @@ Thumbs.db
 vendor
 uploads/*
 !uploads/avatars
+uploads/avatars/*
 !uploads/.gitkeep
-!uploads/avatars.gitkeep
+!uploads/avatars/.gitkeep
