+++ logout.php
@@ -1,0 +1,8 @@
+<?php
+
+require_once('helpers.php');
+
+session_start();
+
+$_SESSION = [];
+header("Location: /");
