+++ templates/main-layout.php
@@ -59,14 +59,18 @@
         </section>
         <section class="authorization">
           <h2 class="visually-hidden">Авторизация</h2>
-          <form class="authorization__form form" action="#" method="post">
+          <form class="authorization__form form" action="index.php" method="post">
             <div class="authorization__input-wrapper form__input-wrapper">
               <input class="authorization__input authorization__input--login form__input" type="text" name="login" placeholder="Логин">
               <svg class="form__input-icon" width="19" height="18">
                 <use xlink:href="#icon-input-user"></use>
               </svg>
               <label class="visually-hidden">Логин</label>
-              <span class="form__error-label form__error-label--login">Неверный логин</span>
+              <span class="form__error-label form__error-label--login">
+                  <?php if (isset($errors)): ?>
+                  <?=$errors['login'];?>
+                  <?php endif; ?>
+              </span>
             </div>
             <div class="authorization__input-wrapper form__input-wrapper">
               <input class="authorization__input authorization__input--password form__input" type="password" name="password" placeholder="Пароль">
@@ -74,7 +78,11 @@
                 <use xlink:href="#icon-input-password"></use>
               </svg>
               <label class="visually-hidden">Пароль</label>
-              <span class="form__error-label">Пароли не совпадают</span>
+              <span class="form__error-label">
+                <?php if (isset($errors)): ?>
+                <?=$errors['password'];?>
+                <?php endif; ?>
+              </span>
             </div>
             <a class="authorization__recovery" href="#">Восстановить пароль</a>
             <button class="authorization__submit button button--main" type="submit">Войти</button>
